Decouple ListLoader from Asteroids and inject ObjectType dependency
