+++ app/services/importer.rb
@@ -1,0 +1,5 @@
+module Importer
+  OBJECT_TYPES = {
+    asteroid: Importer::Asteroids::ObjectLoader
+  }.freeze
+end
